
\name{data_entry}
\alias{TableEntry}
\alias{TableCheck}
\alias{DataCheck}
\alias{Checkall}
\alias{anyErrors}

\title{data entry/error checking functions}
\description{functions used to enter data via excel. TableEntry uses utilsMPIO::XLSForm to return a dataframe with new data; TableCheck looks for new entries in tables, DataCheck looks for errors in tables, Checkall checks all tables using Datacheck, and anyErrors returns TRUE if any table contains an error.}

\examples{


}