
\name{devices}
\alias{DEV_down}
\alias{RFID_up}
\alias{TT_up}
\alias{MSR_up}
\alias{DEV_findfiles}

\title{device upload/download functions}
\description{DEV_down is used to copy RFID/VIDEO files from SD to RAWdir; DEV_findfiles finds new files in RAWdir (not normally RFID or video, but it finds those as well); all the _up functions load data from RAWdir into the logger database.}

\examples{


}